backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pillow==10.2.0
python-dotenv==1.0.1
google-generativeai==0.3.2
requests==2.31.0
chromadb>=0.4.0
langchain>=0.1.0
numpy>=1.24.0
pandas>=2.0.0 
opencv-python>=4.8.0
# Additional AI providers
openai>=1.0.0
anthropic>=0.18.0 